Increase size of kernel stack for purecap kernel.
In CHERI 256 we now use two 64K superpages for the kernel stack,
this is the next available size in the pagemask register so that
a single TLB entry can be used to map the kernel stack.

In CHERI 128 we now use both 16K superpages instead of having a
guard page. The guard page is unnecessary since the stack capability is
bounded.