Fix bus_dmamap_sync_buf for unaligned buffers in CHERI-128.
When dmamap_sync_buf is called for an unaligned buffer, the
portion ouside the buffer that is not aligned to a cache line
must be saved and restored across the dcache invalidation.
In CHERI 128 it may happen that the rest of the cache line contains
a capability, care must be taken to preserve it.
This operation is inherently bad for CHERI because it forces us
to use a root capability to access the rest of the cache line.
For now fix this by enforcing the same alignment between the
temporary buffer and the source buffer, so that memcpy will
preserve capabilities properly.