Initial work on the sys_mmap systemcall for the purecap kernel.
Fix the vm_map_find_min function to return a capability derived from the vm_map root.
Fix vm_mmap_object to return a capability derived from the vm_map root.
Change fileops signature for fo_mmap, this is expected to return a valid capability
the mapping.
For now in sys_mmap we use the same capability source as in the hybrid case, if
no source capability is provided we use the per-thread cheri_mmap_capability. We
should use the one returned by the VM subsystem instead but this makes it harder
to disable bounds enforcement with current sysctls.