Fix the NT_LWPINFO note for CheriABI processes.
Previously, this note was storing the mips64 layout of
'struct ptrace_lwpinfo' instead of the CheriABI layout.

Note: We might consider having an 'imgact_elf64c.c' or some such so
that we could use #ifdef's instead of runtime checks.  (imgact_elf64.c
would then only be for non-CheriABI)  This would reduce the amount of
duplicated code for at least this function.