Allow building setmode.c on Linux
It is needed for bootstrapping since setmode is not provided by glibc.
This change includes a rather ugly hack that declares getumask() as
non-static when building against glibc since the glibc headers include a
non-static getumask() definition but there is no actual implementation.
To workaround this, define a macro GETUMASK_STATIC which expands to static
when not bootstrapping on linux+glibc.