Workaround cyclic dependency when bootstrapping lex and m4
lex needs m4 to build but m4 also depends on lex (which needs m4 to
generate any files). To fix this cyclic dependency we can build a bootstrap
version of m4 (with pre-generated files) then use that to build the real
m4. For lex we also use the pre-gerated files since we would otherwise
need to build awk and sed first (which need lex to build)