Detect st_flags availability when building usr.bin/xinstall
This will not be available when building on Linux. Fortunately the ifdefs
to handle the missing st_flags member already exist so we just need to
detect presence instead of always passing -DHAVE_STRUCT_STAT_ST_FLAGS=1.
This can be done by checking whether the UF_SETTABLE macro is defined.