Avoid use of DDC in global initialization.
Add a crt_init_globals_gdc() which takes a pointer containing the shared
object (which might be the old DDC).  Implement crt_init_globals() in
terms of it for backwards compatability.  Use the new function in _start
and pass the value of AT_PHDR rather than DDC.  This allows DDC to be
cleared at the beginning of _start().