[upstream] Avoid setting empty variables to value 1 in boot_parse_arg()
I noticed this while adding support for parsing the boot commandline in
QEMU MALTA. QEMU will set argv[0] to the value of the -kernel QEMU option
and pass the value of the -append flag as argv[1]. If you start QEMU
without using -append, this will result in an empty string in argv[1].
If we then pass this empty string to boot_parse_arg() will will have an
empty variable with value 1 in the environment.

TODO: I'm not sure whether the check should be added here or in kern_setenv().