Ugly workaround for a compiler bug.
Ensure that the compiler keeps _bitstr live by passing it to a no-op
function that the compiler can't see.  Without this change this
expression:

	_value = ((_curbitstr - _bitstr) * _BITSTR_BITS) + _offset - 1;

was yielding total nonsense resulting in _value > _nbits leading to a
return of -1.  This in turn caused an infinite loop in fork_findpid when
creating early processes (the audit worker or init).