Rework exec_copyin_args() to work for multiple ABIs.
Add a helper function reads a pointer from an array (argv or envv)
and advances the array pointer.  This allows us to remove a decent
chunk of code duplication in the form of freebsd32_exec_copyin_args().

One minor downside of this change is the lost of types on the argv and
envv arrays.