libhelloworld: Don't build with -z notext
The compartment is currently linked as PIE, but the CSU is assembled as
position-dependent and so dla uses an absolute addressing sequence.
Since the .text segment is then writable, LLD erroneously tries to add
64-bit dynamic relative relocations to fix the values up at runtime.
Given that MIPS uses .rel rather than .rela, LLD then writes the addends
for these relocations at the target, which clobbers both the instruction
and the following instruction. LLD should give an error in this case or
emit dynamic HIGHEST/HIGHER/HI16 relocations (LO16 is still constant
since the ABI ensures the load offsets are aligned to 64 KiB), but we
shouldn't even need a writable .text segment.