Handle PT_GNU_RELRO in crt_init_globals phdr parsing
The newest version of LLD will create a separate RWX PT_LOAD for the RELRO
segment so we also need to handle the PT_GNU_RELRO (which will only have
PF_R and not PF_W) in order to add it to the readonly range.
This was causing crashes in webkit init before because the constant
capabilities pointing to the relro segment were being derived from the
readonly capability which did not include that range.