Don't abort when tags are stripped due to an under-aligned target.
While this is finding true positives, there are clear cases where there
could be false positives and we've currently got broken sshd and
dhclient in various cases (both of which are quite hard to debug without
pre-planning).  This in turn is breaking kuya (test) bootstrapping in
jenkins.

Ideally this check should be selectable at runtime (perhaps using an
ifunc for performace).

Additionally, limit printf warnings to once per-process to avoid excessive
spamming.