Fix restoring of PCB registers in purecap kernel.
EPCC is now restored along with PC as in upstream. The scratch capability register used
is parametrized so that we can use a different one in the purecap kernel.
In the purecap kernel MIPS registers are restored before capability registers to allow
the use of scratch capability registers when restoring from PCB.