Fix user exception handler to prevent TLB faults.
The capability for the kernel stack for the current thread can not
be obtained from the pcb capability. This is because the pcb bounds
are strict in the purecap kernel.
In the exception handler we can not access curthread.td_kstack, which
holds the kernel stack capability, without wiring the thread structure
in the TLB. Instead we save a copy of the current kstack capability in
the pcpu data upon swiching.