PLT ABI: Allow creating RTLD function pointers without ELF relocations
In the PLT ABI function pointers need a trampoline that sets $cgp. This does
not work inside RTLD if it is built with the PLT ABI since the initial
relocation process does not have access to globals and dies when it
encounters R_CHERI_CAPABILITY. This could also be fixed by doing another
pass over the relocations but since there are barely any function pointers
in RTLD using a macro is easier.