rtld: Add missing "memory" clobbers to prevent reordering
I was only testing a -O0 rtld so the inline assembly happened to work.
However, at -O2 the compiler is free to reorder instruction before/after
the inline asm (including the restore of $cgp after a function call).
This was causing the rtld $cgp to be re-installed after the inline asm so
the function calls would then crash due to a wrong $cgp.

Using "memory" clobbers should fix this issue but for the two cases the
return the target function pointer it would be much nicer if we could just
return "target+cgp" in $c3/$c4 instead of having to use an on-stack output
argument.

See https://github.com/CTSRD-CHERI/llvm-project/issues/326