[UPSTREAM?] Disable debug output support in RTLD by default
We now install a /libexec/ld-elf-debug.so that can be used instead to debug
issues with the runtime linker. This should improve the performance of
dynamically linked binaries compared to statically linked ones since there
are checks for debug in quite a lot of places.

This avoid the load of a global variable and a branch in many potentially
hot code paths.