Revert "Retire sys/mips/cheri/cheri_test.c."
Bisecting shows that this is causing boot on FPGA to fail once the first
dynamically linked binary is executed. I am not entirely sure why this is
happening but it might be related to the sysctl declarations (since rtld will
use a sysctl to determine the capability size when loading hybrid binaries).

This reverts commit 7806e073af2efcf9913985b1dc78cb8ba62be6e8.