Allow stack pointer to be represented with 14-bit bottom
This requires the stack top to be rounded down by 16 pages instead of just
one page.

We were previously assigning $csp twice during process startup. The second
time seems to be due to a mismerge at some point in the past since it was
not attmempting to round the value at all. Also remove the code to set $sp
since this has not been needed for a long time.