caprevoke: cheritest lib
A test of libcheri_caprevoke's internals.  At a high level:

Allocate a several page arena and fill it with self-referencing
capabilities (base = own address).  Simulate a bump the pointer
allocator revoking after each free; for each, test that the revoker has
cleared exactly the tags requested and no others in the arena.

This includes both tests in a single process and tests in a child
process that has forked.

This test also contains knobs for its paranoia and for verbosity,
should we wish to use it as something of a microbenchmark or, as it was
originally developed, as a debugging aid for libcheri_caprevoke.

Co-authored-by:	Brooks Davis <brooks@one-eyed-alien.net>