caprev vm_machdep: move to specialized authority cap
This is probably the right thing to do, inasmuch as it means we don't
pull a great deal of capability authority into vm_machdep.c.  Nicely,
however, it seems to work around a hardware csetaddr bug while at it.

While here, drop some __cheri_fromptr that aren't what we meant in
printouts.