Add a set of non-destructive and destructive self tests for mbuf-related bounds.  Non-destructive tests use assertions that only fail if the test fails.  Destructive tests throw exceptions if the test passes.  To run a test, set the associated sysctl (in the security.cheri.test tree) to 1.
Destructive tests include violating lower and upper bounds for in-mbuf
data with and without packet headers, and for ordinary clusters.

Further tests are needed for additional cluster sizes.  It would be nice
if we could 'catch' bounds failures and handle them -- e.g., along the
lines of copyin/copyout fault handlers -- to avoid the need to panic the
kernel to check bounds behaviour.