Restrict PCC bounds and remove ability to access memory via kernel DDC.
Make locore, exception and switch routines PIC and simplify assembly
capability initialization. Use explicit relocation macros for capability
and absolute relocation types.
Now the kernel KDC is used to store the kernel globals pointer, the kernel
DDC is NULL and the kernel PCC covers KSEG0 to KSEG3.

Most of the cap initialization is deferred to cheri_init_capabilities now.
KSCRATCH2 is aliased to GPC so that GPC is loaded on exception entry and
restored on exception return. KSCRATCH2 can now be removed an explicitly use GPC.