libcheri_caprevoke: be less of a special snowflake
In particular, build in both hybrid and pure mode.  While hybrid is not
necessarily very useful to us at the moment (as the kernel supports the
syscalls only for cheriabi processes), it has the advantage of being
the "normal" path through the FreeBSD build system.  As a result, our
libraries and, importantly, headers get installed with a minimum of
fuss, as opposed to the earlier attempts at selectively enabling the
build for particular combinations of make variables.