caprevoke: avoid page faults when hunting
When looking for the next possibly-capability-bearing page within a vm
object, there is a simple case in which none of the backing ("shadowed")
objects can store capabilities: the next page in *this* object is the
only possible candidate.  This is readily found by walking the object's
radix tree, with no need for page faults.

Credit to Edward Napierala for doing the hard work; this commit is a
lightly modified version of his original proposal.