libcheri_caprevoke: take two caps for setting bitmaps
libcheri_caprevoke takes on the challenge of preventing double-frees
even across revocation epochs by checking, within a ll/sc block, that
the capability being cited as reason for revocation is still valid.
However, this capability came from the user and so may have incorrect
bounds; as such, a superset capability must be used to describe the
object's shadow bits, but the construction of such a superset replaces
the provenance chain of the user's capability.  Rather than make the
allocator have to work hard to construct a wide capability that is yet
guaranteed to have been revoked if the user's capability has been, just
take both and do the right thing internally.