[subobject-bounds] Fix a sub-object out-of-bounds read in cheritest
We were copying sizeof(reason) bytes from the ccs_testresult_str array
which is twice as large as the array. However, the whole ccsp object is
allocated using a single call to mmap() so there was always enough trailing
data accessible to not cause a bounds violation.
With subobject bounds enabled, the ccsp->ccs_testresult_str array decay
results in a bounded value and therefore crashes in memcpy().

Additionally, this avoids copying TESTRESULT_STR_LEN (1024) bytes of
garbage that will never be used.