caprevoke: respond more correctly? to VM errors
This requires threading error returns further back up the stack and
reworking the outer caprevoke state machine.  Notably, all updates of
the epoch take place at the *end* of the pass.  If an attempt is made to
open a new revocation epoch and this fails, we must not flag that the
epoch has begun.

Not everything is great, since, for example, this attempt makes no
progress when run under gdb, but at least it's not infinite-looping in
the kernel.