caprevoke: try harder to minimize time in thread_single
By default, always do a pass before stopping the world.  If an epoch is
not open, then we open one here; otherwise, this is an incremental pass.
Only then do we stop the world and deal with the kernel hoarders,
register files, and the lingering pages in the VM.

The old behaviour may be obainted with CAPREVOKE_LAST_NO_EARLY for
experimental comparison.