caprevoke: fix revoker interlocking and epoch reports
Knowing what time it is or was is very important and also tricky.
Rework the interlocking at the top of caprevoke() and adds commentary
for users of the revocation subsystem about the intended use of the
values reported.

This fixes at least one subtle bug, wherein we would have told userland
that part of its quarantine had been registered with the kernel by the
current epoch even if the next epoch was in the process of being opened.