caprevoke: tidy vm_do_caprevoke
Don't CAndPerms in the delay slot of the beqz after cexeq.  If the tag
has become unset since we looked at the (fetched earlier) capability
under test, this will throw a CP2 fault; let the compiler insert a nop
into the delay slot, since there's nothing useful we can do there.

Separately, it looks like LLVM can inline this function, so in an effort
to assist it further, expose more of it as C rather than assembler.