[subobject-bounds] Fix bounds violation in localtime
When parsing the timezone file format, this code uses a structure for the header
which is then followed by other structures but are not part of the same struct.
The pointer to the trailing data using is then obtained using
`u->tzhead.tzh_charcnt + sizeof u->tzhead.tzh_charcnt` which sets the bounds
for the buffer to 4 bytes (and it already points to the end so any access will
fault). The fix is simple: There's already a union with a maximum size char
array so we can just add sizeof(tzhead) to index into that buffer instead.