[subobject-bounds] Automatically detect incorrect use of container_of()
This uses the new compiler builtin __builtin_marked_no_subobject_bounds()
to check that the field used in container_of() will not have had bounds
added.
There are probably more copies of the container_of macro in the source tree
that need updating, but for now this allows all of userspace to compile
with subobject bounds.
This might be too aggressive for some uses of container_of() but should
catch fields that have not been annotated.