[subobject-bounds] Annotate "inheritance" in fts() implementation
The FTS structure is embedded as the first member and a pointer to it should
also grant access to the following fields. Since there is only one callsite I
used the __unbounded_addressof() macro instead of adding a general opt-out
annotation to the structure definition.
This fixes the /bin/ls tests in the FreeBSD testsuite.