Correctly initialize iovec lengths in kern_jail().
iovec objects referencing a path, a jailname and a hostname used the same
invalid length, a number of bytes allocated to store copied in strings and
addresses passed to the jail system call.

This bug resulted in invalid memory access in kern_jail_set() that is called
from kern_jail(). When not exploited, the jail system call usually returned
EINVAL or ENAMETOOLONG due to incorrectly terminated strings, according to their
invalid lengths.

Note that FreeBSD does not include this bug. It was introduced with
IOVEC_INIT*() macros in CheriBSD.