caprevoke: slightly optimize capdirty tracking
In the inter-revocation period, there's no real point in differentiating
"historically" and "recently" capdirty, as we're certain to revisit the
"historically" dirty pages anyway when opening the next epoch.
Therefore, when the world is stopped, copy VPO_PASTCAPSTORE up to
PGA_CAPSTORED so that the pmap initially maps pages with the storecap
inhibit clear and doesn't trap their accesses.

This behavior can be toggled with a "caprevoke_last_redirty" sysctl.