caprevoke: maybe not all pages need to be writable
vm_caprevoke_object should not immediately bail if it can't get a
writable mapping of a page.  While it *should* do a read-only probe
before deciding that it does or doesn't need the fault, at the very
least, it can look at the capability dirty flags first.

For whatever reason, hardware has decided that it really needs this
check to avoid failing to revoke, so go ahead and add it as a
placeholder before the real deal (which might wait until post-ASPLOS).