caprevoke: rewrite fullwise, add RO probes
Build one to throw away, as they say.

The initial implementation had a number of shortcomings, which this
attempts to fix.

Not all paths through the earlier system agreed on whether pages should
be considered wired or not.  In particular, after read faults, pages
might have remained wired by mistake.  This may or may not have been the
cause of the KERN_PROTECTION_FAILURE errors we were seeing under the
original implementation.  In this implementation, pages are unwired
after faults wire them and after the vm object lock has been reacquired.

Further, while tempting, the "inversion of control" in the original
vm_caprevoke implementation, whereby the inner page iterator would bail
back to the outer one to find the next page, turns out to not be very
easy to generalize to experimental read-only probing of pages.  In this
more direct formulation, it is comparatively easy to do so, so do.