caprevoke: vm: don't walk off the end of an entry
Bugfix on ab00d6a02cb5ee9513d7b5985ed82947cc1cd1ef

When on the fast path, whereby we know that shadowed objects cannot
store capabilities, we use the topmost object's radix tree to find the
next position.  However, just because the object has a page later
doesn't mean it's actually visible in this mapping of it.

I'm surprised this took this long to show up in testing.  Thanks to
Brett for the report and reproducer.

While here, correct updates to pages_fault_skip; we were failing to add
the intervening pages when we found a next page in range.