caprevoke: redefine "revoked cap"
Mike Roe and David Chisnall point out that programs like to use the tag
to distinguish arms of unions, and so it is important that we not cause
type confusion by swapping a capability for an integer.

Centralize the definition of capability revocation so that it is
consistent everywhere, and change it to *preserve* the tag bit but zero
permissions.  Sealed capabilities are additionally unsealed (but still
have zeroed permissions).  While this may not be the right answer, is's
now articulated in exactly one place.

Catch libcheri_caprevoke's guard against revokedness of free()-d
pointers up to match.  This is a flag day for all caprevoke-ing
allocators: debugging assertions that the tag is stripped must be
changed.