caprevoke: explicitly pass shadows everywhere
Since we're being honest, in that caprev is a property of the vm map, we
shouldn't rely on access to the current process's shadow to check for
revokedness.  At the moment we still do, but the decision is now made
the top rather than in the heart of the mess.