caprevoke: let vm_do_caprevoke decide HASCAPS
We never really cared about tags per se, we wanted to know about
unrevoked capabilities.  Now (since
3f649ec527bc23bf0c026bce280722cbee14fad0) that we aren't stripping tags
when we clear capabilities, the distinction is more pronounced: we don't
want to visit a page if it's got only revoked capabilities (and/or data)
on it.  So, when sweeping a page in vm_caprevoke_page, let its helper
function, vm_do_caprevoke, decide whether a capability counts as
dirtying the page.

The interesting upshot of this is that we might have pages in the system
which are (historically, recently, MD) capclean and yet have tagged caps
on them and have the (tag-stripping) load inhibit cleared!  It's fine if
the user loads and manipulates a revoked capability; if they store it,
we'll take splash damage, but that seems like it will be less likely.