WIP: caprevoke: more precisely consider capdirty
An earlier commit (244640fb4fa81520c18a10ae2d74f86a8dd435a9) noted that
pages with VPO_PASTCAPSTORE asserted will certainly be visited in the
next epoch opening, and so we would like to avoid taking spurious faults
to re-dirty them in the inter-epoch period.  Asserting PGA_CAPSTORED is
even a reasonable way to cause the pmap to map the page with the
capstore inhibit clear.

However, this results in pmap_sync_capdirty marking all pages held in
the TLB(s) as dirty.  We are interested in a subset: those pages that
are capdirty since we last looked at them (if we have).  Add another
page flag, VPO_CAPREVSECOND to indicate that the revoker may have
cleared PGA_CAPSTORED during the opening pass.  This is still an
over-estimate, but a better one, and it still keeps us down to one time
that we need IPIs.  (Nevermind that we actually do two, one to
thread_single and one to invalidate TLBs.  In principle these could be
fused.)

It is, admittedly, a little uncomfortable to be taking up three fairly
precious bits, but at least two of them are in vm_page_t's oflags field,
which need not be atomic and so could be easily expanded.