Add basic support for stack traces in the pure capability kernel.
Now we can run with KDB enabled. DDB is not tested yet.
The stack unwinder takes advantage of the fact that function boundaries are
easily determined by the base and length of code capabilities.
The unwinding logic is similar to the legacy MIPS one, which finds the
offsets of FP and RA in the stack frame from the store instructions that
save them.
This could be probably be improved by using actual DWARF unwinding info but
for now this seems stable enough.