[upstream?] Reduce output from getserv_test
Before the getservbyport test printed 142927 lines of output. This seems to
confuse kyua when generating a junit xml report and we were getting
corrupted xml missing the <system-out> start tag. Additionally, all this
output slows down the test unnecessarily

Also put the "structures are not equal:" output behind #ifdef DEBUG since
it the test only fails if another condition is true. This avoids lots of
output in the