libmalloc_simple: Add semi-passive revocation.
Memory is queued for revocation on free() and revocation is triggered on
allocation when memory would have to be allocated by mmap() and
appropriate memory is in quarantine.

We add a second set of buffers for allocations that have been painted for
revocation.  Rather than doing a full revocation when we exceed
MAX_QUARANTINE bytes in the quarantine buffer, that triggers painting
the shadow bitmap for the allocations and updating the current epoch.
Actually revoking is defered until MAX_PAINTED bytes for allocation have
been painted (e.g. that much will be recovered).

Further, revocation will not be inititiated unless at least one
free'd allocation of the size we want will be reclaimed.

Features from Brooks, bugs from nwf.

Co-authored-by: Nathaniel Wesley Filardo <nfilardo@microsoft.com>