Fix sub-object buffer overread in libarchives tar parser
When parsing an ascii number it was unconditionally loading the next byte at the
end of the loop. This is not exploitable since the next character is always in
the same structure and the value is never used. However, it does crash with
sub-object bounds and the loop should be restructured.