Squash the CPU_CHERI in sysproto.h.
This one truly is CHERI MIPS specific, and in fact specific to 256-bit
CHERI.  Change _PAD() to use syscallarg_t instead of register_t and
use it on all but CHERI 256 on MIPS.  For CHERI 256 simplify the
previous macros and re-arrange them to reduce the diff with upstream.

Note that I have only tested 128, not 256.