Set bounds on function pointers in the GOT during relocation.
All function pointers are bounded to a subset of the root kernel
code capability.
Some assembly functions for exception handlers and context switching are
manually set to size 0, which leaves PCC unbounded.
This works around the presence of branches across function boundaries for now.
The trap() handler is modified to return the full EPCC to install instead of
just the PC offset, this is required for proper handling of page fault
during copyin/copyout operations.