Align cmsghdr data to pointer size.
This was previously avoided as it does not make sense to pass capabilities.
The pure capability kernel needs pointer alignment for SCM_RIGHTS messages as the
file descriptors integer indices are converted to/from pointers to the file structure in
unp_internalize and unp_externalize.
It is possible to twak unp_internalize to align the new cmsghdr data buffer so that we can
align to pointers, however it seems simpler to allow everything to be aligned since we can
not rely on misalignment to prevent users sending capabilities anyway.