Fix cmsghdr realignment to kernel pointer size for sendmsg and recvmsg.
Add realignment of the msg_control entries for compat freebsd64 in sendmsg.
Realignment of msg_control for the hybrid kernel/purecap userspace should not be required
if we always align to capability size in the hybrid kernel.
Minor refactoring of freebsd64 recvmsg to avoid changing user pointers when writing to the
output msghdr.