Switch pointers in the default ABI to use __capability.
This is another step towards having the default ABI be CheriABI.
Note that this is (mostly) the minimal set of changes to get the
kernel compiling again.  Not all system calls have been audited
to see if copyin/copyout need to be replaced with *cap versions
instead.

In general the changes consist of removing __USER_CAP*() for
system call arguments and using EXPLICIT_USER_ACCESS in files
with system calls that use copy*() to use the *_c versions by
default.

In some cases it seemed simpler to move forward with annotating
pointer fields in structures shared with userland with
__kerncap rather than adding additional __USER_CAP*() invocations.

For example, mount_argsu() was changed to take a capability for a
user pointer for the nmount case, but that then impacted vfs_cmount
implementations, so I changed vfs_cmount to take a capability for
the user pointer and used __kerncap in the FS-specific argument
structures.

For the NFS I did choose to add some __USER_CAP*() as the rabbit
hole was getting quite deep, so NFS will have to be revisted to
plumb capabilities in the shared structures in the future.

A few places such as NFS also do some horrific things where they
use a structure shared with userland that contains pointers and
overwrite the user pointers with kernel pointers at "outer" layers
and pass in the modified structure to "inner" layers which know
they are receiving kernel pointers instead of user pointers.  This
required constructing in-kernel capabilities.  This would all
fall out in a purecap kernel.

sys_sigqueue's change is a temporary hack until the sigval types
are cleaned up.

For __umtx_op_nwake_private, I defined BATCH_SIZE for the default
ABI to use 1K of stack space and retired BATCH_SIZE_C.