Rebuild kstack capability to include the pcb area for the kstack cache.
In the purecap kernel, we split the kstack area and use two capabilities to
access the actual kstack and the pcb.
When the thread is destroyed, the kstack is cached to be reused whenever
a new thread is created. This assumes that the cached kstack pointer is
enough to hold the stack and pcb regions, which is not true anymore.

This rederives the original kstack allocation when the thread is destroyed.
Ideally we should have a way to cache the pcb and kstack pointers separately.