Add an initial sysentvec and ELF activator for purecap for MIPS.
This adds a new debug.use_cheriabi sysctl node.  If the node is true
(default), COMPAT_CHERIABI is used to run purecap binaries.  If the
node is false, the "native" activator is used.  The native activator
isn't close to working yet, but some of the really broken things are
marked with TODO so that they can be fixed.