Don't build stack-protector code for CHERI purecap
This should never be used when we have capabilities. The compiler will emit
an error when you try to enable stack-protector in purecap mode so these
symbols should never be referenced. As we also don't need FORTIFY in purecap
mode (and FreeBSD libc doesn't include it anyway), none of the run-time
symbols provided by libc/secure are needed.