Stop using both pc and pcc in trap handling
While working on sentry capabilities I noticed that having the split
between $pc and $pcc is awkward since we have to ensure that we don't
setoffset on $epcc (by writing to EPC) if $epcc is a sentry capability.

The approach I have chosen here is to use trapframe->pc everywhere.
I also changed trapframe->pc to an opaque pointer type to get compiler
errors rather than the implicit conversions that void* allows.

When CHERI is supported we use the slot for $pcc instead and replace
the 64-bit pc slot with padding.
In order to avoid increasing the size of this patch, both trapframe->pcc
and trapframe->pc are currently supported using anonymous union.
I will remove this union in a follow-up commit.

In order to update the assembly files and hopefully catch all uses of EPC
I decided to use #pragma GCC poison to avoid using MIPS_COP_0_EXC_PC and
the PC macro when CHERI is supported.

Unfortunately this commit is huge since I couldn't break it down into smaller
changes that still compile/run.