Adjust subobject bounds on mld6 hop-by-hop extension header.
The ip6_pktopts for ip6_output from mld6 is intended to point to the full
struct mld_raopt through the ip6_pktopts.ip6po_hbh field.
The ip6po_hbh field is a pointer to ip6_hbh, which is just the header of
the struct mld_raopt. This ends up using the wrong subobject bounds when the
ip6po_hbh field is assigned in mld_init().

This sets the proper bounds on ip6po_hbh, however it seems desirable to
make it more explicit that the ip6po_hbh field it's an ip6_hbh plus
some extra data.