Work around compiler bug reordering subobject bounds operations.
In tcp_input, the address of next_hop6->sin6_addr is taken in a conditional block.
The bounds for the pointer are computed before the branch check, causing a length
violation.
This workaround removes the subobject bounds for this specific case.

This might be related to https://github.com/CTSRD-CHERI/llvm-project/issues/359