Allow reservations to be merged in non-CheriABI processes.
While there might be some value in padding some maps in hybrid code,
it's easy enough for code like libcheri to do it's own padding.  If we
implemented cmmap we might want to make reservation mergability
a per-entry property to support mixed mmap/cmmap code, but this works
for now.