libcompat: Add WITH_COMPAT_CHERIABI to build purecap libraries.
This removes the custom code derived from the old freebsd32 support
into the Makefile.libcompat framework.  While we don't think this is a
good long-term model of purecap support, it is helpful during bringup
to be able to run a native/hybrid userspace with a limited set of
purecap applications.

One addition to the Makefile.libcompat code is building internal
libraries to we can build an entierly purecap world on top of the
libcompat setup.  It's not clear if this is something we want to
upstream, but it might be with another knob to control it.