Change uintptr interpretation in the purecap kernel from offset to address.
This removes all uses of the ptr_to_va macro.
Wherever the virtual address interpretation needs to be made explicit, ptr_to_va
is replaced by a cast to vaddr_t or vm_offset_t.
This happens when we do arithmetic between two intptr_t or pointer and intptr_t.