crt1: Don't zero a large Elf_Auxinfo array
The latest compiler will recognize this and replace the loop with a call
to memset(): https://github.com/CTSRD-CHERI/llvm-project/commit/c87fe2f1f2a9e303c183586510c0c7db3c86e512
However, we can't call memset() before crt_init_globals has completed since
the captable is still uninitialized.

To fix this replace the complete copy of auxargs with a single iteration
that reads only the values that we use inside _start().
As the code now includes an if-else chain we have to compile the file with
-fno-jump-table. Using jump tables require loading a capability from
the captable and this cannot be done before crt_init_globals().