Temporarily disable all DMA support on the purecap DE4 kernel.
This is due to a type mismatch in bus_dma_segment_t for bus_addr_t, which
is used to hold both bus addresses and virtual addresses or device virtual
addresses in an IOMMU world. This difference needs to be made explicit and
probably we need assertions on tag invariants that must be checked when
necessary.