Create an MI file to hold routines for building caps for exec.
Move some of the subroutines I recently added next to the MIPS
exec_setregs() for CheriABI along with another helper for setting
the sigcode capability to sys/cheri/cheri_exec.c.  Use this to
add untested purecap support in the RISC-V exec_setregs() as well
as setting cra for the purecap sendsig().