Don't set pcc, idc, or csp in hybridabi_sendsig().
pcc is immediately overwritten in freebsd64_sendsig() after this
function returns to point to the signal handler.  idc and csp
should probably not be written as global ddc.

We should perhaps be clearing all capability registers to a known state
rather than leaving whatever register values were in place at the time
of the signal, but the signal handler can already get to all of those
values via the saved values in the signal frame anyway.

This does leave user ddc in a signal handler set to the default ddc
instead of whatever value was present at the time the signal is
delivered.