Rework the freebsd64 mcontext_t for CHERI on RISC-V.
Don't try to save just pcc and ddc in the padding.  Instead, go with
the MIPS-style getcontextx() even with the merged register file where
the base structure contains a virtual address pointer to a separate
copy of the capability register file on the stack.

For now, always assume that the capability register file is
authoritative when provided except that the integer PC is preferred.
This might not be ideal, but I'm a bit leery of trying to guess when a
signal handler might have modified an integer register in the
mcontext_t, etc.