Allow re-registering conames if they refer to SCBs owned by dead threads.
Modify the check for existing conames to unseal the c_value capability to the switcher control block and check if the td has been set to NULL, as happens when a thread exits. If it has been, we remove this coname from the list, and allow the calling thread to register the name itself. If it has not been set to NULL, behaviour remains the same.

This allows coaccept(2)-using programs to work when they are opportunistically colocated in the address space an instance of the same once resided in.