Fix further ioctl problems.
One of them was quite serious: the definition of DTRACEIOC_EPROBE had dtrace_eprobedesc_t as an argument (size=56), while it should have been a pointer(size=8 or 16).
In this way, in sys_generic.c user_ioctl(), 56 bytes were copied in from the address in the ioctl parameter, instead of only 8 (or 16). With cheri, this gave an error,
because the ioctl parameter was a capability with len 0x10: as a consequence there was a memory error.