Fix problem with uarg in many dtrace structs.
uint64_t were used in many places to keep pointers to user arguments.
I had to change them to kuintcap_t in the kernel to make it work, and
change a bit of allignment checks.
Another problem was in dtrace_dof_copyin: even if the struct to copy
from userspace to kernelspace doesn't have pointers, a field in it
determines the size of a further following copyin from a buffer. There,
there are capabilities. Initially, I was copying it with copyin_c. copyin_c
strips the validity tag, so copyincap should have been used instead.